fix issue AO maps rotation
01-Rotation -
There is an issue with rotation of imported models. Previous models ("Posed") has been importing correctly, but after import the "Posed-Plus" model, the rotation is wrong. I attach previous model ("Posed") & PosedPlus for comparison, as both models import same & correctly into Blender. There is a difference in the models scales and axes but all Posed Plus models have same as in example. Please correct it so it works 🙂

02-Diffusion - Diffusion in hair material imports wrong texture. It should be AO.
The Mix mode must be "Normal".

03-AOTex - Color profile on AO textures must be sRGB/Default